+++ example/lib/swagger_generated_code/client_index.dart
@@ -1,0 +1,2 @@
+export 'example_swagger.swagger.dart' show ExampleSwagger;
+export 'pets_api.swagger.dart' show PetsApi;