+++ lib/src/code_generators/swagger_models_generator.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 
-import 'package:collection/src/iterable_extensions.dart';
+import 'package:collection/collection.dart';
 import 'package:swagger_dart_code_generator/src/code_generators/constants.dart';
 import 'package:swagger_dart_code_generator/src/code_generators/swagger_enums_generator.dart';
 import 'package:swagger_dart_code_generator/src/code_generators/v3/swagger_enums_generator_v3.dart';
@@ -219,9 +219,9 @@ abstract class SwaggerModelsGenerator {
 
     final listEnums = getAllListEnumNames(dartCode);
 
-    listEnums.forEach((listEnum) {
+    for (var listEnum in listEnums) {
       results = results.replaceAll(' $listEnum ', ' List<$listEnum> ');
-    });
+    }
 
     return results;
   }
@@ -283,9 +283,10 @@ abstract class SwaggerModelsGenerator {
     final words = parameterName.split('\$');
 
     final result = words
-        .map((e) => e.pascalCase
+        .map((e) => e
             .split(RegExp(r'\W+|\_'))
-            .map((String str) => str.capitalize)
+            .mapIndexed(
+                (int index, String str) => index == 0 ? str : str.capitalize)
             .join())
         .join('\$');
 
@@ -369,11 +370,11 @@ abstract class SwaggerModelsGenerator {
 
     jsonKey = jsonKey.camelCase;
 
-    forbiddenCharacters.forEach((String element) {
+    for (var element in forbiddenCharacters) {
       if (jsonKey.startsWith(element)) {
         jsonKey = '\$forbiddenFieldName';
       }
-    });
+    }
 
     if (jsonKey.startsWith(RegExp('[0-9]')) ||
         exceptionWords.contains(jsonKey)) {
@@ -1175,6 +1176,9 @@ $generatedProperties
 \tstatic const fromJsonFactory = _\$${validatedClassName}FromJson;
 \tstatic const toJsonFactory = _\$${validatedClassName}ToJson;
 \tMap<String, dynamic> toJson() => _\$${validatedClassName}ToJson(this);
+
+@override
+String toString() => jsonEncode(this);
 
 $equalsOverride