+++ test/generator_tests/enums_generator_test.dart
@@ -23,7 +23,7 @@ void main() {
 
     test('Should generate enum from request parameter', () {
       final result =
-          generator.generate(request_with_enum_in_parameter, 'test_file');
+          generator.generate(requestWithRnumInParameter, 'test_file');
 
       expect(result, contains('enum V3OrderOrderIdStatePutOrderStateRequest'));
     });
@@ -38,8 +38,11 @@ void main() {
     test('Should generate enum values', () {
       final values = <String>['file_sup'];
       const output = "\t@JsonValue('file_sup')\n\tfileSup";
-      final result =
-          generator.getEnumValuesContent(enumValues: values, isInteger: false);
+      final result = generator.getEnumValuesContent(
+        enumValues: values,
+        isInteger: false,
+        enumValuesNames: [],
+      );
 
       expect(result, contains(output));
     });
@@ -68,13 +71,13 @@ void main() {
 
     test('Should remove numbers at beginning if it is key word', () {
       final result =
-          SwaggerEnumsGenerator.getEnumNamesFromRequests(request_with_enum);
+          SwaggerEnumsGenerator.getEnumNamesFromRequests(requestWithEnum);
       expect(result[0], equals('PetsPetIdItemsGetContentType'));
     });
 
     test('Should remove numbers at beginning if it is key word', () {
       final result = SwaggerEnumsGenerator.getEnumNamesFromRequests(
-          request_with_list_of_enum_in_parameter);
+          requestWithListOfEnumInParameter);
       expect(result[0], equals('V3OrderOrderIdStatePutOrderStateRequest'));
     });
   });